K8s refactor (#9)
* bumping version to 0.8.0

* update version to 0.9.0

* smore refactoring for k8s and tests

* pad support
* server-side dry-run

* k8s refactoring and clean up

* k8s refactoring
* deploy src support
* render k8s scripts
* more tests
* temp target cleanup

+++ main/render/deploy.sh
@@ -43,14 +43,16 @@ function doRender {
     export _RENDER_ERRORS=""
     for fname in $config_files
     do
-      rm -rf $_DEPLOY_ERROR_DIR/config_error.log
-      echo -e "------------------------------------------------------\nrendering $fname"
-      actualRender "$fname"
-      if [ -f $_DEPLOY_ERROR_DIR/config_error.log ]; then
-        echo "failed to render: $TEMPLATE_DIR/$fname" >> $_DEPLOY_ERROR_DIR/render_error.log
-        cat $_DEPLOY_ERROR_DIR/config_error.log | sed -e "s/^/  /g" >> $_DEPLOY_ERROR_DIR/render_error.log
+      if [ "$fname" != "*.srd" ]; then
+        rm -rf $_DEPLOY_ERROR_DIR/config_error.log
+        echo -e "------------------------------------------------------\nrendering $fname"
+        actualRender "$fname"
+        if [ -f $_DEPLOY_ERROR_DIR/config_error.log ]; then
+          echo "failed to render: $TEMPLATE_DIR/$fname" >> $_DEPLOY_ERROR_DIR/render_error.log
+          cat $_DEPLOY_ERROR_DIR/config_error.log | sed -e "s/^/  /g" >> $_DEPLOY_ERROR_DIR/render_error.log
+        fi
+        echo -e "end $fname\n------------------------------------------------------"
       fi
-      echo -e "end $fname\n------------------------------------------------------"
     done
     cd $_curdir
     if [ -f $_DEPLOY_ERROR_DIR/render_error.log ]; then
